requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
SQLAlchemy==2.0.31
psycopg[binary]==3.2.1
python-dotenv==1.0.1